+++ cypress/fixtures/utils.ts
@@ -16,6 +16,15 @@ export const getIframe = (): Cypress.Chainable<JQuery<HTMLBodyElement>> => {
     .then((body) => cy.wrap(body as HTMLBodyElement));
 };
 
+export const openEditLink = () => {
+  return getIframe()
+    .findByTestId('cf-ui-popover-content')
+    .should('exist')
+    .findByLabelText('Edit link')
+    .should('exist')
+    .click({ force: true });
+};
+
 export const getIframeWindow = () => {
   return cy
     .get('#storybook-preview-iframe')
@@ -23,3 +32,8 @@ export const getIframeWindow = () => {
     .should('not.be.empty')
     .then(cy.wrap);
 };
+
+// copied from the 'is-hotkey' library we use for RichText shortcuts
+const IS_MAC =
+  typeof window != 'undefined' && /Mac|iPod|iPhone|iPad/.test(window.navigator.platform);
+export const mod = IS_MAC ? 'meta' : 'control';