+++ packages/rich-text/src/plugins/Hyperlink/components/ResourceHyperlink.tsx
@@ -1,46 +1,54 @@
 import * as React from 'react';
 
 import { FieldAppSDK, Link } from '@contentful/app-sdk';
-import { TextLink } from '@contentful/f36-components';
+import { Text } from '@contentful/f36-components';
 
 import { useContentfulEditor } from '../../../ContentfulEditorProvider';
 import { findNodePath, isChildPath } from '../../../internal/queries';
 import { Element, RenderElementProps } from '../../../internal/types';
 import { useSdkContext } from '../../../SdkProvider';
-import { handleCopyLink, handleEditLink, handleRemoveLink } from './linkHandlers';
+import { useLinkTracking } from '../../links-tracking';
+import { useResourceEntityInfo } from '../useResourceEntityInfo';
+import { handleEditLink, handleRemoveLink } from './linkHandlers';
 import { LinkPopover } from './LinkPopover';
 import { styles } from './styles';
 
-type HyperlinkElementProps = {
+export type ResourceHyperlinkProps = {
   element: Element & {
     data: {
-      uri?: string;
       target: {
         sys: {
-          id: string;
-          linkType: 'Entry' | 'Asset';
-          type: 'Link';
+          urn: string;
+          linkType: 'Contentful:Entry';
+          type: 'ResourceLink';
         };
       };
     };
   };
-  target?: Link;
-  onEntityFetchComplete?: VoidFunction;
+  target: Link;
+  attributes: Pick<RenderElementProps, 'attributes'>;
   children: Pick<RenderElementProps, 'children'>;
 };
 
-export function UrlHyperlink(props: HyperlinkElementProps) {
+export function ResourceHyperlink(props: ResourceHyperlinkProps) {
   const editor = useContentfulEditor();
   const sdk: FieldAppSDK = useSdkContext();
   const focus = editor.selection?.focus;
-  const uri = props.element.data?.uri;
+  const { target } = props.element.data;
+  const { onEntityFetchComplete } = useLinkTracking();
   const pathToElement = findNodePath(editor, props.element);
   const isLinkFocused = pathToElement && focus && isChildPath(focus.path, pathToElement);
 
+  const tooltipContent = useResourceEntityInfo({ target, onEntityFetchComplete });
+
+  if (!target) {
+    return null;
+  }
+
   const popoverText = (
-    <TextLink className={styles.openLink} href={uri} rel="noopener noreferrer" target="_blank">
-      {uri}
-    </TextLink>
+    <Text fontColor="blue600" fontWeight="fontWeightMedium" className={styles.openLink}>
+      {tooltipContent}
+    </Text>
   );
 
   return (
@@ -48,17 +56,18 @@ export function UrlHyperlink(props: HyperlinkElementProps) {
       isLinkFocused={isLinkFocused}
       handleEditLink={() => handleEditLink(editor, sdk, pathToElement)}
       handleRemoveLink={() => handleRemoveLink(editor)}
-      handleCopyLink={() => handleCopyLink(uri)}
       popoverText={popoverText}
     >
-      <TextLink
+      <Text
         testId="cf-ui-text-link"
-        href={uri}
-        onClick={(e) => e.preventDefault()}
+        fontColor="blue600"
+        fontWeight="fontWeightMedium"
         className={styles.hyperlink}
+        data-resource-link-type={target.sys.linkType}
+        data-resource-link-urn={target.sys.urn}
       >
         {props.children}
-      </TextLink>
+      </Text>
     </LinkPopover>
   );
 }