+++ packages/rich-text/src/plugins/Hyperlink/components/styles.ts
@@ -9,11 +9,27 @@ export const styles = {
       fontSize: 'inherit !important',
     },
   }),
+  iconButton: css({
+    padding: `${tokens.spacing2Xs} ${tokens.spacingXs}`,
+  }),
+  openLink: css({
+    display: 'inline-block',
+    marginLeft: tokens.spacingXs,
+    marginRight: tokens.spacingXs,
+    maxWidth: '22ch',
+    whiteSpace: 'nowrap',
+    overflow: 'hidden',
+    textOverflow: 'ellipsis',
+  }),
+  popover: css({
+    zIndex: tokens.zIndexModal,
+  }),
   hyperlink: css({
     fontSize: 'inherit !important',
     display: 'inline !important',
     '&:hover': {
       fill: tokens.gray900,
+      textDecoration: 'none',
     },
     '&:focus': {
       fill: tokens.gray900,